fixup! [open][neva_upstremable][v8_snapshot] V8 Snapshot implementation
Release Notes:
Fixup for V8 Snapshot implementation

:Detailed Notes:
There was a crash when playing a video in
Enyo2sampler -> Moonstone -> VideoPlayerInline.

The crash has caused by keeping CHECK(IsSane()). However, the CHECK was
upstream code and it was already removed in chromium68 though, the
"V8 Snapshot implementation" kept the code. So this patch removes the CHECK()
to be aligned with upstream.

Additionally this patch modifies the type of some arguments and variables.

:Testing Performed:
Check if video plays well on Enyo2sampler -> Moonstone -> VideoPlayerInline.

:QA Notes:
Test Cases: GT-25869

:Issue Addressed:
[PLAT-28331] Provide mksnapshot error logs
[PLAT-38986] Migrate all V8 patches to new V8 branch
[PLAT-66066] Chromium 2018Q4 upgrade: migrate "v8" patchset - V8 Snapshot
[PLAT-69523] Renderer crash when playing media

Change-Id: I9773f10db6eee23e25e00835274cd4e0ab51a9e6
Reviewed-on: http://gpro.lgsvl.com/222855
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Gyuyoung Kim <gyuyoung.kim@lge.com>
Tested-by: Gyuyoung Kim <gyuyoung.kim@lge.com>
Reviewed-by: Jose Dapena <jose.dapena@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>