[open][downstream][dst] Fix date problem with extremely short term of DST
:Release Notes:
Fix date problem with extremely short term of DST

:Detailed Notes:
In certain platforms DST delta may be set with extremely short
period; an hour. But V8 date cache regard that least term of DST is 20
days (in case of Egypt). So if (custom) short period of DST might be
placed in the middle of boundary where V8 think that period is no-DST
zone.
With this CL, V8 would check DST with delta of 59 min, not 19
days. Of course, there might be less cache hits and more cache
updates. Still, an hour can be thought as an acceptable one.

:Testing Performed:
1) Set a DST period of 1 hour
2) Put clock just before the hour DST starts
3) Wait till clock gets to that hour
4) The hour should move to DST immediately.

:QA Notes:
Test-cases: N/A
Depends on: http://gpro.lgsvl.com/231001

:Issues Addressed:
[PLAT-65881]

Change-Id: I93a2f3b5b115dfb9cca44e820bd1846a611b277d
Originally-Reviewed-on: https://gpro.lgsvl.com/#/c/171110/
Reviewed-on: http://gpro.lgsvl.com/231000
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Jose Dapena <jose.dapena@lge.com>
Tested-by: Jose Dapena <jose.dapena@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Reviewed-by: Jani Hautakangas <jani.hautakangas@lge.com>